refactor(#1471): split blocks/registry.py + delete legacy IO finder per ADR-047 (Phase 3 D2)
Decompose src/scistudio/blocks/registry.py (1708 LOC) into the
src/scistudio/blocks/registry/ sub-package using the Path D pattern,
and delete the pre-ADR-043 legacy IO finder API in the same PR.

Sub-package layout (all <750 LOC, ADR-047 §C9 compliant):
- registry/__init__.py — BlockRegistry class + bindings, BlockSpec,
  5 error classes, helper re-exports
- registry/_scan.py     — builtin/tier1/tier2/monorepo scan +
  _register_spec + capability cross-validation
- registry/_capability.py — list_format_capabilities, find_*_capability,
  matching primitives, _resolve_*_class helpers
- registry/_spec.py     — _spec_from_class, ADR-030 MRO merge,
  ADR-038 distribution-version resolution

Deleted (pre-ADR-043 transitional surface):
- find_loader / find_saver / find_io_blocks_for_type
- _find_io_block, _class_accepts_dtype, _best_specificity,
  _matching_capabilities_for_legacy_io, _resolve_legacy_capability_class

Caller migrations:
- blocks/code/validation.py — dead legacy fallback branch removed
- blocks/io/materialisation.py:_format_supported_savers — rewritten on
  list_format_capabilities; output shape preserved
- 10 governance docs frontmatter (ADR-008/009/025/028/029/030/036/037/043
  and docs/specs/adr-043-io-format-capability-registry.md) updated to
  governs.files: src/scistudio/blocks/registry/**

Tests:
- DELETE tests/blocks/test_registry_find.py (deleted API)
- DELETE 3 legacy fallback tests in test_block_registry_capabilities.py
- NEW tests/blocks/test_registry_package_layout.py — AST guard +
  public import surface preservation + legacy-absent assertions

Tooling: scripts/check_god_files.py — registry waiver removed.

Closes #1471
