+++ .workflow/records/1471-registry-decomp.json
@@ -1,0 +1,191 @@
+{
+  "admin_labels": [
+    {
+      "applied_at": null,
+      "applied_by": null,
+      "name": "admin-approved:core-change"
+    }
+  ],
+  "amendments": [
+    {
+      "approved_by": null,
+      "exclude": [],
+      "include": [
+        "docs/adr/ADR-008.md",
+        "docs/adr/ADR-009.md",
+        "docs/adr/ADR-025.md",
+        "docs/adr/ADR-028.md",
+        "docs/adr/ADR-029.md",
+        "docs/adr/ADR-030.md",
+        "docs/adr/ADR-036.md",
+        "docs/adr/ADR-037.md",
+        "docs/adr/ADR-043.md",
+        "docs/specs/adr-043-io-format-capability-registry.md"
+      ],
+      "reason": "Closure findings required updating governs.files frontmatter on 10 ADRs/specs that previously referenced the deleted src/scistudio/blocks/registry.py to point at the new src/scistudio/blocks/registry/** sub-package glob."
+    }
+  ],
+  "branch": "refactor/issue-1471/registry-decomp",
+  "changed_test_paths": [
+    "tests/blocks/test_registry_package_layout.py",
+    "tests/blocks/test_block_registry_capabilities.py"
+  ],
+  "check_results": [
+    {
+      "command_or_tool": "ruff check src/scistudio/blocks/registry/ src/scistudio/blocks/code/validation.py src/scistudio/blocks/io/materialisation.py tests/blocks/test_block_registry_capabilities.py tests/blocks/test_registry_package_layout.py",
+      "exit_code": 0,
+      "name": "ruff check (scoped)",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "ruff format --check src/scistudio/blocks/registry/ src/scistudio/blocks/code/validation.py src/scistudio/blocks/io/materialisation.py tests/blocks/test_block_registry_capabilities.py tests/blocks/test_registry_package_layout.py scripts/check_god_files.py",
+      "exit_code": 0,
+      "name": "ruff format (scoped)",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "pytest tests/blocks/test_block_registry_capabilities.py tests/blocks/test_registry_package_layout.py tests/workflow/ tests/contracts/ tests/blocks/code/ tests/blocks/io/ --timeout=60 --no-cov",
+      "exit_code": 0,
+      "name": "pytest scoped suites",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "python scripts/check_god_files.py --enforce",
+      "exit_code": 1,
+      "name": "check_god_files (registry waiver removed)",
+      "output_path": null,
+      "status": "unknown",
+      "summary": {},
+      "timestamp": null
+    }
+  ],
+  "commit": null,
+  "docs_landing": {
+    "adr": {
+      "paths": [
+        "docs/adr/ADR-047.md"
+      ]
+    },
+    "changelog": {
+      "not_applicable": true,
+      "rationale": "Structural refactor + legacy API removal with zero observable behavior change. Phase 3 D2 of umbrella 1427."
+    },
+    "checklist": {
+      "not_applicable": true,
+      "rationale": "Single-PR implementation."
+    },
+    "docs": {
+      "paths": [
+        "docs/adr/ADR-047.md"
+      ]
+    },
+    "spec": {
+      "not_applicable": true,
+      "rationale": "ADR-047 authoritative."
+    }
+  },
+  "full_audit": {
+    "blocks_merge": false,
+    "command": "python -m scistudio.qa.audit.full_audit",
+    "exit_code": 0,
+    "known_debt": [],
+    "output_path": "audit-out.json",
+    "status": "pass",
+    "summary": {},
+    "unclassified_failures": []
+  },
+  "governance_touch": false,
+  "issues": [
+    {
+      "close_in_pr": true,
+      "followup_rationale": null,
+      "number": 1471,
+      "url": "https://github.com/zjzcpj/SciStudio/issues/1471"
+    }
+  ],
+  "owner_directive": "Implement ADR-047 registry decomposition + legacy IO finder removal",
+  "planned_files": [],
+  "pull_request": null,
+  "record_path": ".workflow/records/1471-registry-decomp.json",
+  "required_checks": [],
+  "schema_version": "1",
+  "scope": {
+    "exclude": [],
+    "include": [
+      "src/scistudio/blocks/registry/**",
+      "src/scistudio/blocks/registry.py",
+      "src/scistudio/blocks/code/validation.py",
+      "src/scistudio/blocks/io/materialisation.py",
+      "tests/blocks/test_block_registry_capabilities.py",
+      "tests/blocks/test_registry_find.py",
+      "tests/blocks/test_registry_package_layout.py",
+      "scripts/check_god_files.py",
+      "docs/adr/ADR-047.md",
+      ".workflow/records/1471-registry-decomp.json"
+    ]
+  },
+  "sentrux": {
+    "command_or_tool": "mcp__plugin_sentrux_sentrux__scan + check_rules + health",
+    "mode": "free-tier",
+    "name": "sentrux.free_tier",
+    "output_path": null,
+    "pro_required": false,
+    "quality_signal": 4333.0,
+    "rules_checked": 3,
+    "status": "pass",
+    "summary": {},
+    "thresholds": {
+      "quality_signal": "4000"
+    },
+    "total_rules_defined": 15
+  },
+  "stages": [
+    {
+      "completed_at": null,
+      "stage": "scope_and_issue",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "plan",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "implement",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": "2026-05-22T23:36:05.987238+00:00",
+      "stage": "update_docs",
+      "status": "done",
+      "summary": "docs_landing populated via JSON edit (gate_record docs is destructive); ADR-047 + 10 governance docs governs.files frontmatter updated for sub-package layout."
+    },
+    {
+      "completed_at": null,
+      "stage": "test_and_checks",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "commit_and_submit_pr",
+      "status": "pending",
+      "summary": null
+    }
+  ],
+  "task_id": "1471-registry-decomp",
+  "task_kind": "refactor"
+}
