+++ docs/adr/ADR-008.md
@@ -23,7 +23,7 @@ governs:
     - scistudio.types
   files:
     - docs/adr/ADR-008.md
-    - src/scistudio/blocks/registry.py
+    - src/scistudio/blocks/registry/**
     - src/scistudio/core/types/registry.py
     - pyproject.toml
     - tests/architecture/test_registries.py
