+++ docs/adr/ADR-009.md
@@ -23,7 +23,7 @@ governs:
   entry_points: []
   files:
     - docs/adr/ADR-009.md
-    - src/scistudio/blocks/registry.py
+    - src/scistudio/blocks/registry/**
     - src/scistudio/core/types/registry.py
     - tests/architecture/test_registries.py
   excludes: []
