+++ docs/adr/ADR-025.md
@@ -30,7 +30,7 @@ governs:
     - docs/adr/ADR-025.md
     - src/scistudio/blocks/base/package_info.py
     - src/scistudio/blocks/base/__init__.py
-    - src/scistudio/blocks/registry.py
+    - src/scistudio/blocks/registry/**
     - src/scistudio/core/types/registry.py
     - pyproject.toml
     - tests/blocks/test_registry.py
