+++ docs/adr/ADR-028.md
@@ -53,7 +53,7 @@ governs:
     - src/scistudio/blocks/io/loaders/**
     - src/scistudio/blocks/io/savers/**
     - src/scistudio/blocks/base/block.py
-    - src/scistudio/blocks/registry.py
+    - src/scistudio/blocks/registry/**
     - src/scistudio/api/schemas.py
     - frontend/src/utils/computeEffectivePorts.ts
     - tests/blocks/test_dynamic_ports.py
