+++ docs/adr/ADR-029.md
@@ -40,7 +40,7 @@ governs:
   files:
     - docs/adr/ADR-029.md
     - src/scistudio/blocks/base/block.py
-    - src/scistudio/blocks/registry.py
+    - src/scistudio/blocks/registry/**
     - src/scistudio/api/schemas.py
     - src/scistudio/workflow/validator.py
     - src/scistudio/blocks/ai/ai_block.py
