+++ docs/adr/ADR-030.md
@@ -26,7 +26,7 @@ governs:
     - scistudio.blocks
   files:
     - docs/adr/ADR-030.md
-    - src/scistudio/blocks/registry.py
+    - src/scistudio/blocks/registry/**
     - src/scistudio/blocks/io/io_block.py
     - src/scistudio/blocks/app/app_block.py
     - src/scistudio/workflow/serializer.py
