+++ docs/adr/ADR-036.md
@@ -28,7 +28,7 @@ governs:
     - docs/adr/ADR-036.md
     - src/scistudio/api/routes/projects.py
     - src/scistudio/api/routes/lint.py
-    - src/scistudio/blocks/registry.py
+    - src/scistudio/blocks/registry/**
     - frontend/src/components/**
   excludes: []
 tests:
