+++ docs/adr/ADR-037.md
@@ -28,7 +28,7 @@ governs:
     - docs/adr/ADR-037.md
     - pyproject.toml
     - src/scistudio/cli/install.py
-    - src/scistudio/blocks/registry.py
+    - src/scistudio/blocks/registry/**
     - src/scistudio/core/versioning/git_binary.py
   excludes: []
 tests:
