+++ docs/adr/ADR-043.md
@@ -51,7 +51,7 @@ governs:
     - src/scistudio/blocks/io/**
     - src/scistudio/blocks/io/capabilities.py
     - src/scistudio/blocks/io/simple_io.py
-    - src/scistudio/blocks/registry.py
+    - src/scistudio/blocks/registry/**
     - src/scistudio/blocks/io/materialisation.py
     - src/scistudio/blocks/app/**
     - src/scistudio/workflow/validator.py
