+++ docs/adr/ADR-047.md
@@ -27,7 +27,10 @@ governs:
   entry_points: []
   files:
     - docs/adr/ADR-047.md
-    - src/scistudio/blocks/registry.py
+    - src/scistudio/blocks/registry/**
+    - src/scistudio/blocks/code/validation.py
+    - src/scistudio/blocks/io/materialisation.py
+    - scripts/check_god_files.py
   excludes:
     - tests/**
     - docs/audit/**
